enhanced favorites action
Added additional parameter 'uri' to favorites action. Goal is to return a unique identifier for favorites. Use case is to toggle with just one button from an ESP8266 through radion stations listed as favorites in Sonos. The amount of data returned by the 'detailed' parameter was too heavy for the ESP8266 memory.